+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added helper functions for keeping track of cell/range dependencies: getCellPrecedents and getCellDependents. (#441)
+- 11 text functions CONCATENATE, SPLIT, LEN, TRIM, PROPER, CLEAN, REPT, RIGHT, LEFT, SEARCH, FIND.
+- helper methods for keeping track of cell/range dependencies: getCellPrecedents and getCellDependents. (#441)
 - 4 financial functions FV, PMT, PPMT, IPMT.
 
 ## [0.1.3] - 2020-07-21