+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- 11 text functions CONCATENATE, SPLIT, LEN, TRIM, PROPER, CLEAN, REPT, RIGHT, LEFT, SEARCH, FIND.
+- 9 text functions LEN, TRIM, PROPER, CLEAN, REPT, RIGHT, LEFT, SEARCH, FIND.
 - helper methods for keeping track of cell/range dependencies: getCellPrecedents and getCellDependents. (#441)
 - 4 financial functions FV, PMT, PPMT, IPMT.